agent/pyproject.toml
[project]
name = "building-permit-agent"
version = "0.1.0"
description = "AI agent for reviewing building plans against California Building Standards Code."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.135.1",
    "google-adk[a2a,otel-gcp]>=1.26.0",
    "google-cloud-aiplatform>=1.140.0",
    "google-cloud-documentai>=3.11.0",
    "google-cloud-storage>=3.9.0",
    "google-genai>=1.66.0",
    "httpx>=0.28.1",
    "opentelemetry-api>=1.38.0",
    "opentelemetry-exporter-gcp-trace>=1.11.0",
    "opentelemetry-instrumentation-fastapi>=0.59b0",
    "opentelemetry-instrumentation-httpx>=0.59b0",
    "opentelemetry-sdk>=1.38.0",
    "pydantic>=2.12.5",
    "pytest>=9.0.2",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.22",
    "reportlab>=4.4.10",
    "uvicorn>=0.41.0",
]
